refactor: create lecture utils_test.py
